XOrgConfCreator: Try to deal with the Intel Gen11+ gamma table mess.
The situation on Intel graphics hardware:

- Good old intel-ddx is basically a minimally maintained zombie, which
  is not used by default on any modern distribution, at least not since
  Ubuntu 20.04-LTS. The driver only supports hardware older than Gen-10,
  so Gen-9.x Skylake/Kabylake is the end of the road.

- All Gen11/12/13/... Icelake/Tigerlake/Alderlake/DG-1/DG-2/... hardware
  only works with the modesetting-ddx which is well maintained and the
  default driver for Intel. Since X-Server 21 ~ Ubuntu 22.04-LTS it does
  support GAMMA_LUT high precision gamma tables, as the intel-ddx did,
  so enables > 8 bpc precision output for 10 bpc depth 30 deep color
  framebuffers etc.

- However, since Gen-12 Tigerlake/DG-1/Alderlake-S hardware had hardware
  design bugs in GAMMA_LUT handling which weren't worked around at server
  21 release time, the modesetting ddx by default disables GAMMA_LUT's on
  gpu's with more than 4096 slots -- the only culprits being Gen11 Icelake
  and the broken Gen12. All older Intel gpus and recent Gen13 gpu's do not
  have the hardware bugs and use 1024 slot tables.

- The Linux 5.15 kernel and later do have driver workarounds for the broken
  hardware, so use of GAMMA_LUT is considered safe until proven otherwise
  for modern distributions like Ubuntu 22.04 with Linux 5.15 or later. But
  the modesetting-ddx of X-Server 21 does not know this.

Therefore we must use a manual override for Intel gpu's under modesetting
ddx for deep color framebuffer modes if we want to actually display 10 bpc
framebuffers with at least 10 bpc output precision on Gen 11 and Gen 12.

This commit implements the override to force-enable use of GAMMA_LUT's
on Intel gpu's with modesetting-ddx in 10 bpc color mode. In 8 bpc mode
or on non-Intel gpus, the decision is left to the modesetting-ddx if and
when GAMMA_LUT's are used. Atm. only Intel is known to need special baby
sitting - knock on wood!

I can't actually test if this all works out properly as all my Intel
gpus are Gen 9.5 - too old.