AssertOpenGL: Update Matlab troubleshooting for Ubuntu 20.04-LTS build system.
Since we switched the build system for the Matlab mex files from
Ubuntu 18.04-LTS to Ubuntu 20.04-LTS, and install the libdc1394.so.25
devel package for builds, the Screen.mexa64 file also hard-depends on
libdc1394.so.25 instead of the .22 version.

This means it just works(tm) on Ubuntu 20.10+, but on Ubuntu 20.04-LTS
libdc1394.so.25 is only installed iff a matlab/octave-psychtoolbox-3
package from Neurodebian/Debian/Ubuntu pulls in the needed library or
if sudo apt build-dep psychtoolbox-3 pulled in the dependency.

If none of this happened and users just get PTB via Download/UpdatePsychtoolbox
then libdc1394.so.25 might be missing on Ubuntu 20.04-LTS, leading to
load-time link failure for Screen().

Update the AssertOpenGL workaround to symlink to libdc1394.so.22 in this
case to resolve the issue with a hack.