Screen: Add interop support infrastructure for OpenXR / PsychOpenXRCore.
For interop between Screen() rendered stimuli and future PsychOpenXRCore,
aka an OpenXR runtime for VR/AR/MR/XR applications, we need to render
and post-process stimulus images as usual via OpenGL in Screen, and then
pass those to the OpenXR runtimes OpenGL backend for consumption and
display by the XR compositor.

For this, the OpenXR runtime needs access to a suitable OpenGL rendering
context and its configuration, so it can use that context to create OpenGL
compatible XR swapchains, export OpenGL renderable textures from it, and
we can import those into Screen as finalizedFBO rendering targets. The
context must get created by Screen() in a manner compatible with Screen's
rendering, but the context must be suitable for exclusive use by the
OpenXR runtime, potentially on runtime-private threads. So we need a
dedicated OpenGL context.

Our solution for now is to (ab-)use the OpenGL rendering context normally
created by Screen() for use by Screen's background parallel flipperThread.
This thread is normally used for async flips (ie. Screen('AsyncFlipBegin/
Check/End')), and for our own homegrown frame-sequential stereo
implementation (stereomode 11 or our stereomode 1 fallback), and our own
VRR fine-grained stimulus timing implementation. However, apart from the
async-flip case, the flipper thread does not have any use in a XR stimulation
scenario, as it would be the job of the XR compositor to do whatever is
needed to present XR stimuli properly.

Therefore we define a new Screen('Openwindow', ..., specialFlags); flag
kPsychDontUseFlipperThread. This flag, when passed in from the OpenXR
driver, will prevent use of the flipperThread - trying to use it will
cause an error abort. As the OpenGL glswapcontext for the flipper thread
is always and unconditionally created during Screen('OpenWindow'), but
now not actually needed for the unused flipper thread, we can use it
instead for OpenXR interop, as this context gets already properly setup
for these kind of activities.

A new 'infotype' 9 query type is added to Screen('GetWindowInfo', win, 9)
that allows to query all required information about the glswapcontext
created by Screen(), e.g., the OpenGL context handle, drawable, pixelformat,
and device context, so the info can be exported from Screen() after
OpenWindow, and imported into PsychOpenXRCore for OpenGL interop setup with
the OpenXR runtime.

Flow:

1. Screen('OpenWindow') with kPsychDontUseFlipperThread specialflag, so
   the flipperthread's OpenGL resources get created, but the thread won't
   get used.

2. info = Screen('GetWindowInfo', win, 9) to query properties of the
   OpenGL context and resources created in step 1.

3. Use info to pass those into PsychOpenXRCore for OpenGL-OpenXR interop
   setup for OpenGL rendering/processing by Screen() and consumption and
   XR composition by the OpenXR compositor.

-> This commit adds the needed flags and api and help texts and setup
   code for doing this on Linux/X11/GLX and Windows/Win32. macOS does
   not support OpenXR at all atm., so this is Linux/Windows only atm.