Screen('GetWindowInfo'): Fix new infoType 9 on Windows.
We can't expose pixelformatObject on Windows as a pointer or uint64, as it is a
whole struct, not an opaque pointer to a struct. However, we don't need it anyway
for OpenGL <-> OpenXR interop on Windows, only the DeviceContext (HDC) and
OpenGLContext (HGLRC) matter, so just return a constant zero.