windowsmakeit64_twisty: Add build support for future OpenXR driver on Windows.
Add a new plugin type 16 for building PsychOpenXRCore.mex[w64], an Octave
and Matlab driver for use of OpenXR for VR/AR/MR/XR under MS-Windows.
On Windows, we statically link the OpenXR loader openxr_loader.lib, which was built from
source from Khronos OpenXR-SDK with MSVC 2019, into our mex file, whereas on other
systems like Linux, dynamic linking against an OS provided loader is used. This follows
recommendations by the SDK makers.

The driver itself is not ready yet, but lets add buid system already to simplify maintenance
and later integration, as well as ongoing development.

Note that building under Matlab R2022a was successful and the driver works the same on
Windows so far, as it did on Linux. Building under Octave-6.4 currently fails with many
undefined reference errors. Untested suspicion is that the openxr_loader.lib built with
MSVC 2019 is not suitable for static linking with Octave's MinGW 64 toolchain, and that
we might need a dedicated build of the openxr_loader.lib built under MinGW64 for
compatibility. That would be tedious but not insurmountable. For now I'll just carry on
under Matlab only, as this is not the highest priority problem to tackle right now.

-> Basic test of basic PsychOpenXRCore.mexw64 driver under Windows-10 21H2 with the OculusXR
runtime worked.