PsychOpenXR: Warn if Float16Display precision requested, but unsupported by runtime.
This happens, e.g., on the OculusXR runtime, and leads to workarounds which make it work,
but at output precision loss, e.g., fp16 -> rgba8 unorm, and extra overhead due to Screen
internal blit due to buffer unsharing.

Curiously the OculusXR runtime supports RGB16F, but I am not sure if that is framebuffer
renderable, and obviously no alpha channel in that case. RGB10 or RGB10A2 is also not
supported.