PsychOpenXRCore: Add instance event processing and session state management.
First rough implementation, will need refinement or refactoring.
processXREvents() polls, dequeues and processes XR instance events,
and updates the tracking of the current session state accordingly,
as well as info if session is active, and if the frame presentation
loop should be running to sync with the XR compositor.

Currently we call this from 'EndFrameRender' after each swapchain image
release, but we should probably call it from many / most driver entry
points, or potentially from a future background processing thread.