PsychOpenXR[Core]: Add minimal test implementation of frame presentation.
This is far from what it needs to be, but just minimal code to test the
PTB rendering + presentation + OpenXR compositor pipeline, ie. a proof
of concept to give us something to work with.

Will probably need to run on a background thread multi-threaded later
on, with latching PTB rendered swapchain images at the right time.

-> Test with VRHMDDemo.m and ImagingStereoDemo.m for pure monoscopic and
   stereoscopic drawing (ie. not tracked, not full 3D) under Monado
   emulator at least gives a reasonably looking mono/stereo picture in
   the debug window for the simulated HMD. Further tests pending...