PsychOpenXRCore: Don't PsychErrorExitMsg in functions called from Screen().
'GetNextTextureHandle', 'EndFrameRender', and 'PresentFrame' are typically
called from within Screen's imaging pipeline, so in case of errors they
must not use PsychErrorExitMsg(), as that implies a Screen('CloseAll')
which shuts down and releases the onscreen window that may be needed during
further error handling and pipeline execution, resulting in a crash!

Only print error messages, or return error status codes to caller, but
don't error abort.