PsychVRHMD/PsychOculusVR1('GetEyePose'): Drop some guarantee for an irrelevant feature.
The advertised ability of 'GetEyePose' to also return local/global head
pose as a byproduct was always a lie on PsychOculusVR1, it never did that,
and would also be difficult to implement on the new PsychOpenXR driver, so
it can not be guaranteed across all VR drivers and should be removed from
the PsychVRHMD docs for 'GetEyePose'.

However, not only was this feature never used by any of our demos or tests,
there also ain't any meaningful use-case for it. Ergo, drop it from
PsychVRHMD and all drivers where this isn't easy to do.