PsychOpenXR[Core]: Allow switch/enable of 3D projection view display.
If the basicTask is now one that requires true perspective projected
OpenGL 3D rendering, regardless if tracked or non-tracked, setup will
pass in the new use3DMode flag as 1/true, and initial setup will attach
the projectionLayer/projectionViews as submitLayers for frame submission,
instead of the quadViews layer used for pure 2D drawing and display (ie.,
monoscopic, or stereoscopic Screen() 2D drawing commands like in
ImagingStereoDemo.m). We also initialize and update projectionViews
pose and 3D field-of-view fov information accordingly, so the compositor
knows what it deals with in terms of imaging and scene composition.