PsychOpenXRCore: Implement 'GetStaticRenderParameters'.
This computes suitable OpenGL GL_PROJECTION_MATRIX perspective projection
matrices from the tracked eye views field of view fov and user specified
(or default) clip near/far planes, for perspective correct 3D rendering
with OpenGL.

To build the matrices, we reimplement glFrustum() based on older code
written by myself in glFrustum.m for OpenGL-ES, which in turn was written
based on the official OpenGL specs math and matrix definitions for
glFrustum(). The glFrustum implementation is not 1-to-1, but modified
to use tan() angle mapping of field of view, and to be invariant in its
projection volume shape to the zNear clip plane position. See comments
at top of code.

-> With this and a few hacks, I can get VRHMDDemo1.m to display a
   reasonably looking teapot, so far tested on Linux with MonadoXR in
   emulator mode and also with real Rift-CV1 HMD. Results look good.