PsychOpenXR: Improve MSAA setup for optimal quality vs. performance tradeoff.
If the passed in multisample parameter from PsychImaging('Openwindow', ..., multisample, ...); is
empty, ie. usercode wants the driver to make an intelligent choice, then:

- Request a MSAA level of zero, ie. no anti-aliasing, if OpenXR runtime can not do MSAA itself.

- Request MSAA with the OpenXR runtime recommended "optimal" number of MSAA samples if
  the runtime can do MSAA internal to the XR compositor. This should give a very good quality
  vs. performance/resource consumption tradeoff.

If usercode provided an explicit MSAA setting, then don't override it, but respect it. If the level
is within the MSAA range supported by the OpenXR runtime, then set the imagingmode flag
kPsychSinkIsMSAACapable, so Screen will use MSAA with zero-copy processing if possible and
leave MSAA ops to the OpenXR runtime for best performance and lowest overhead. If the MSAA
requested exceeds the OpenXR capabilities, we let Screen itself do the MSAA and MSAA resolve
internally, at the cost of one extra framebuffer copy from drawBuffer to inputBuffer for the MSAA
resolve, and operate the OpenXR swapchains in standard non-MSAA single-sample mode.

Note that current MonadoXR on Linux is not MSAA capable. OculusXR on Windows is MSAA
capable up to 4x MSAA, but the Oculus OpenXR runtime still recommends only non-MSAA aka
single-sample operation. So in practice, leaving multisample as [] default will not use MSAA on
either of these runtimes, explicit multisample > 0 will alway use internal Screen MSAA on Linux
with Monado, and internal  MSAA beyond 4x on Windows + Oculus VR runtime.