PsychPortAudio: Add new AM modulator flag kPortAudioAMModulatorNeutralIsZero.
This new flag 256 (= kPortAudioAMModulatorNeutralIsZero) selects the
neutral output gain modulation value for a logically stopped AM modulator
as zero, instead of the default value of one.

This means a stopped modulator will not act as if no modulator is
attached at all - which is the current and past default behaviour for
a neutral 1x multiplicator gain. Instead, the 0x multiplicator gain for
a stopped modulator will multiply all sound with 0 and thereby create
silence / mute the sound output on the modulated channels. One can use
this to "gate off" sound that plays on a playback device/channel.

Limitations:

Note that this only works if the modulator slave is attached to another
playback slave device (the most common case), but not if the modulator
is attached directly to the master device. This is a principal limitation,
but luckily only for the more uncommon use case.