osxmakeit: Minor backwards compatibility fix for Screen.
tinyexr.cc was not compiled with mac os version min setting 10.11, so the
built Screen.mexmaci64 for Matlab had a symbol dependency that is only
availabe on macOS 10.15+, ergo Screen did not work on macOS 10.14 or
earlier anymore. Specifically ____chkstk_darwin is the offender.

-> Fix this for Screen.mexmaci64, so it probably works again on pre-10.15

However, our latest rebuild of PsychPortAudio for Octave and Matlab on
macOS Catalina also encoded the same problematic dependency into the
libportaudio_osx64.a library, so PsychPortAudio is broken for either pre-10.15
or pre-10.14, definitely for 10.13 and earlier. I don't intend to fix this anytime
soon without a support membership payment or similar sponsoring, because
we can't act as a charity forever.

Also, so far only 1 person complained about 10.13 brokeness, so maybe nobody
uses those old systems anymore and everyone is now on 10.15+.