PsychPortAudio: Allow 'sysdefault' as auto-selected default audio device.
Match to exclude 'default', but let 'sysdefault' pass the filter, instead
of filtering out anything with 'default' in its name, e.g., 'sysdefault'.

This seems to do no harm when testing under Ubuntu 20.04-LTS on a laptop
with builtin sound chip - same default device selection.

As an improvement it allows auto-selection of 'sysdefault' as the sound
device of choice on the RaspberryPi 400 under RPi OS 11. This is important,
because the old filtering logic rejected sysdefault and any other viable
sound device as default device, falling back to 'default' as device of
choice, which is driven by PulseAudio - and thereby leads to application
hang, as the PulseAudio sound server is suspended by default during a
PsychPortAudio session.

We didn't have this problem on the RPi-2B, as it had a proper sound chip
for output via headphone/line-out port. I assume also no problems on other
RPi 3/4 models. The RPi 400 however does not have speaker/line/headphone
outputs. Its only audio output is via HDMI video output to a connected
audio capable HDMI display, and other problems and peculiarities of sound
output to HDMI on the RPi 400 caused silence or hangs, and least when
testing on RaspberryPi OS 10 / Debian 10 Buster. Since upgrading to RPi
OS 11 / Debian 11 / Bullseye, at least the hdmi and sysdefault devices
work with PsychPortAudio, so we should allow PsychPortAudio to select the
'sysdefault' audio device as default sound output device for a good
plug & play experience. Playback via GStreamer movie playback always
worked, as that went/goes through PulseAudio which takes care of trouble.