Screen: Rework drawBufferFBO unsharing: Trigger for format not on white-list.
Instead of listing all formats that may trigger an drawBufferFBO->
finalizedFBO unshare, rather first have a white-list of know-how-to-handle
formats, then unshare-triggers.

This way, any format unknown to Screen will always trigger an unshare,
as it is not on the white-list.

Of those formats that are known in principle, specific checks are made
to see if an unshare is still needed, e.g., due to range or precision
issues, ie. because they would degrade value range or precision wrt.
what user code wanted for the drawBufferFBO.

-> This should make sure that if Screen follows the "better safe than
sorry" strategy of unsharing if it doesn't know what to do with a format,
because a slight performance degradation is more acceptable than wrong
results or precision loss.