Screen('GetWindowInfo'): Always return a GPULastFrameRenderTime > 0.
The value GPULastFrameRenderTime == 0 is meant to mean "query result
not yet available" and trigger a retry by user scripts later on, or
potential immediately in a tight polling loop. The same result could
also happen though if the query result was available, but happened to
be 0 nsecs, because the Screen main OpenGL context did not have any
gpu work to do in the query period. This ambiguity could cause hangs
due to being stuck in an infinite GPULastFrameRenderTime == 0 polling
loop, e.g., as observed with VRHMDDemo1.m. We can't fix the api ambiguity,
but we can fix the problem: If the query results in 0 nsecs used gpu time,
bump the result up to 1 nsec of used gpu time, so we always get a non-zero
result on successful query and clients get unstuck.

Note: The specific trigger for this bug fix was VRHMDDemo1.m getting
stuck on the fast AMD Raven Ridge gpu if zero-copy interop mode was
used with the OpenXR MonadoXR testing, as the imaging pipeline did not
have anything to do and that was the only use for the Screen main
OpenGL context in that demo.

This points to one problem: Screen('GetWindowInfo') gpu time queries
are of limited use if the imaging pipeline or Screen 2D drawing commands
are not used much/at all, and all rendering is mostly/exclusively done
by userspace OpenGL rendering, ie. between Screen('Begin/EndOpenGL')
calls. During such OpenGL rendering via MOGL glXXX() functions or external
rendering, e.g., by Horde3D, the Screen OpenGL main context is not doing
anything, and 'GetWindowInfo' only measures gpu execution on that main
context. Iow. we only get reporting of imaging pipeline overhead, unless
we mostly use Screen() 2D drawing commands.

In that sense, using this gpu rendertime query in VRHMDDemo1.m or other
true 3D rendering demos is a bit misguided, and may mislead users into
using the wrong approach for time measurement.