Screen: Allow zero-copy for SRGB8A8 if substituting for RGBA8.
The GL_SRGB8_ALPHA8 format is able to fulfill the unorm range and
precision requirements of a GL_RGBA8 drawBufferFBO. Range and alpha
encoding/precision is identical. Color precision of SRGB8 is in practice
a tad higher than of RGB8, as the same amount of 8 bits gets distributed
in a perceptually more effective way for human perceivers and ~gamma 2.2
display hardware. Therefore allow to share drawBufferFBO and finalizedFBO
if the user selected a standard 8 bpc virtual framebuffer and the external
consumer wants SRGB8.

This mostly helps performance for VR/AR runtimes which want so consumer
SRGB8 content, to avoid an extra framebuffer copy, e.g., OculusVR1 and
OpenXR (both OculusXR and MonadoXR, likely others).

It is conceivable that this could even substitute for RGB10_A2 without
practical precision loss, as 8 bit SRGB non-linear may well equate 10
bit linear on suitable display hardware. But lets be rather conservative
at the beginning. Also, Screen does not use RGB10_A2 internally currently,
so no harm done any way.

-> In practice, the PsychOculusVRCore1 driver for OculusVR-1.x on Windows
probably always injected SRGB8 textures into our drawBufferFBO and this
was not rejected due to missing checks, with no harm done. Same likely
for OculusXR runtime on Windows. MonadoXR on Linux is more explicit about
this, and needs opt-in by the PsychOpenXRCore driver, so lets make this
ability to share SRGB8 explicit in the code to avoid confusion or second
guessing in the future.