PsychOpenXR[Core]: Add haptic feedback support via 'HapticPulse' function.
Supports the same functionality as the PsychOculusVR1 and PsychOpenHMDVR
drivers, ie. a sine-wave vibration of selectable duration, frequency and
amplitude. As usual, runtimes + controllers may have various limits wrt.
frequency and duration, and may clamp or discretize accordingly. E.g.,
the Oculus touch controllers can only do two discrete frequencies, 160 Hz
for both models, and 320 Hz for the Rift-CV1 vs. 500 Hz for the Rift-S controllers.
Maximum pulse duration on at least the Rift-CV1 touch controllers seems to be
limited to a maximum of 2.5 seconds, apparently a firmware limit.

One problem is that there ain't no way to detect at session creation
if a give HW supports haptic feedback, so we can only hard-code the
hapticFeedbackSupported flag to 1 for "always supported in principle".
Also no 100% reliable way to detect at runtime if a 'HapticPulse' call
will actually trigger some haptic feedback. We can detect and warn if
the application does not have input focus, which prevents haptic feedback,
but even if the app has input focus, we don't get feedback if any actual
hw is connected with haptic feedback support. One could in principle
query the currently active interaction profile for the haptic output
path and try to derive from that if output is likely generated, but that's it.

Currently implemented is hand controller haptic support, e.g., as successfully
tested for the Oculus Rift CV-1 1st gen touch controllers under Windows with
the OculusXR runtime. Also supported is the Microsoft XBox controller and
equivalent gamepad controllers with interaction profile "microsoft/xbox_controller"

Note that at least under Windows+OculusXR, the runtime seems to latch haptic
setting at 'HapticPulse' execution time, but not actually perform the hardware
programming to start the pulse for real. That only seems to be happening during
the next Screen('Flip') and its associated OpenXR operatings, like xrWaitFrame,
xrBeginFrame or xrEndFrame. This is technically not according to spec, but it is
what it is. The whole haptic support of runtimes seems to be quite shaky and full
of unpleasant surprises.