PsychPortAudio(): Also use dynamic binding of PaUtil_SetDebugPrintFunction() on Windows.
We only used to dynamically bind PaUtil_SetDebugPrintFunction() on Linux at runtime, but use
load-time linking for Octave on Windows and delay-load linking for Matlab on Windows. This
causes hard failure (PsychPortAudio load failure on Octave, PsychPortAudio crash on Matlab) if
the used portaudio DLL does not export the PaUtil_SetDebugPrintFunction() function. While our
own self-built and bundled portaudio DLL does support PaUtil_SetDebugPrintFunction(), the same
apparently is not true for various other alternative portaudio DLL's, so they could not be used as
drop-in replacements (in PsychtoolboxRoot folder) for our DLL. E.g., the libportaudio.dll bundled
with recent releases of Matlab - as tested with R2022a - does not support that function and would
cause trouble.

This fixes it, and the Matlab provided libportaudio.dll can be used after renaming it to
portaudio_x64.dll and copying it into the PsychtoolboxRoot() folder, as tested with the
DLL bundled with R2022a on both Matlab R2022a and Octave 6.4.

When we are at it, improve diagnostic in case of dll load failure and other minor stuff.