PsychPortAudio: Enable minimal tweaks to allow setup of 3rd party provided ASIO backends.
This does not in any way reenable ASIO support on Windows, but if a 3rd party custom portaudio dll
would expose an ASIO backend, because that 3rd party would have done all the needed work and
would carry/handle all the related license obligations and liabilities, this would allow the driver to
do a more fine-tuned job of setting up. We only use PortAudio public API for this, and in no way
depend on or use or link to the proprietary ASIO SDK, so don't need a license for this. Only the
distributor of an ASIO enabled portaudio dll would need such a license with all its obligations.

Note that Psychtoolbox does not and will not ever ship a portaudio dll with ASIO support again.
This just would allow to take a little bit easier advantage of a 3rd party plugin the user might install
on their machine as an override to our portaudio dll. Specifically it does the following:

- Select the ASIO backend for lowlatency operation (reqlatencyclass >= 1) if supported, instead
of WASAPI.
- Use a suggestedLatency of 1 msecs for reqlatencyclass > 1 and 5 msecs for reqlatencyclass 1 as
a better default than 0 msecs normally used.
- Some help text updates in GetDevices?