PsychPortAudio: Improve help text and diagnostics for logical -> physical channel mapping.
Print true logical -> physical channel assignment if 'selectchannels' is specified in PPA('Open') under
MS-Windows with a Windows native sound backend, e.g., WASAPI: Due to design, physical output
channels are selected via a bitmask, so ordering info of which logical channels goes with a physical
channel is lost. Assignment is always in ascending order, ie. effectively 'selectchannels' gets sorted
ascending. This is a limitation compared to other supported sound backends like macOS/CoreAudio.

Also add diagnostic output if trying to use a backend without support, e.g., MME, DirectSound,
WDMKS. Only WASAPI is supported atm. on the Windows native backend side.

Extend help text for 'selectchannels' in PsychPortAudio('Open') and point out the numerous
limitations of channel mapping on Windows, as well as the loss of cross-platform compatibility
and even compat between different sound hardware if this option is used.

Btw. while testing shows that the code works, the chosen channel mapping at least for a mono
1 channel output to headphones on a Realtek onboard audio chip had no effect. Mono sound
was not output to the selected channel 0 or 1 (left or right) exclusively, but to both headphone
speakers. This both in shared WASAPI mode and exclusive WASAPI mode. Could be a limitation
of the onboard sound chip, or a limitation when using headphones instead of real speakers.