Base/OSX: Switch only supported/tested macOS version from 10.15 to 12.
No more development or testing on 10.15.7 Catalina, now that it has been
wiped from my drive. We keep macosx-version-min at 10.11 for the time
being, so PTB may still work back to 10.11, but no guarantees, and I don't
care if it breaks on older systems.

The only available system for development, maintenance and testing for
now is 12.6 Monterey, with XCode 14, macOS 12 SDK, Intel processor only.

Clean up support status messages, when we are at it.