SetStereoSideBySideParameters(): Add basic RemapMouse() support.
Need explicit remapping if offsets and scaling are non-default.
This was a missing piece since day 1 in late 2012.

Limitation: Does not take previously defined inverseWarpMap's into
account. Stuff setup by Panelfitter works with this, but taks like
'FlipHorizontal/Vertical' or 'GeometryCorrection' won't, as the
mappings won't concatenate. Not sure if there is a good solution for
this that is easy to implement. Baby steps...