SetStereoSideBySideParameters(): Make it play nicely with other imaging tasks.
Make sure we can concatenate dynamic stereo changes with previously
statically defined spatial transformations like 'FlipHorizontal',
'FlipVertical', etc. Detect this and implement a remap() function
which concatenates the old/static and new/stereo transform.

Note the for-loops in the remap() function are rather inefficient,
some clever vectorization would do them good. Atm. runtimes can be
on the order of hundreds of msecs on Matlab, hundred-thousands on
Octave which doesn't have a JIT compiler to on-the-fly optimize!
Matlab clearly wins here. Optimizations are left for another day...