Screen: Make 'Hookfunction' - 'GetDisplayBufferTextures' work in more cases.
Allow the query function whenever the imaging pipeline is fully enabled,
to query color buffer backing textures of the finalizedFBO's, which are
always defined when the pipeline is on. Even in case of no actual FBO's,
a pseudo-FBO is defined, which describes the system OpenGL framebuffer of
the onscreen window, ie. texture handles are zero.

So far this was restricted to when finalized FBO sinks were requested, but
there are other potential future use cases.