Screen: Fix PsychImaging task 'MirrorDisplayTo2ndOutputHead' for most cases.
Turns out that this task only worked for trivial configurations without
use of the panelfitter, MSAA, image processing or other complexities.

Fix this, and in the process get rid of some M-Code to execute in the
right finalizer blit chain and save one framebuffer blit caused by that.

This still uses the same bounce-buffer trick, using a finalizedFBO to
make a backup copy of the onscreen window framebuffer / final stimulus
image to that FBO, then switching to the OpenGL context of the slave window
before blitting that finalizedFBO back into the "other" finalizedFBO
which represents the system OpenGL backbuffer of the slave window, thereby
copying the image from main window to slave window for mirroring.

However, while the mechanics worked for dual-window stereomode 10, in
a merged stereomode mirroring failed, as did in standard mono mode,
because the final stimulus image in all these modes is routed to the
finalizedFBO[0], so if that one is the system framebuffer then pipeline
flow optimization often optimized "too much" away and we ended with the
final stimulus image in the system backbuffer, from which we could not
copy to the bounce buffer. Change this by using finalizedFBO[0] as
initial bounce buffer instead of finalizedFBO[1] as done previously. This
way finalizedFBO[0] is always a real FBO, the final stimulus image ends
up there, and then we can blit from finalizedFBO[0] to finalizedFBO[1],
which is initialized to be the system framebuffer / OpenGL window backbuffer
both in the main window with imaging pipeline, and in the slave window.
Iow. blitting to finalizedFBO[1] actually blits to the respective OpenGL
backbuffer of each window -> Stuff works in mono mode (0) and all merged
stereo modes (2-9) which don't need separate image streams, but merge
everything into a single stream. Also tested with simultaneous use of
nothing, panel-fitter, MSAA, image processing ('FlipHorizontal' etc.) and
all combinations of these.

Note that for dual-window stereomode 10 we need to retain the old method
with finalizedFBO[1] as bounce buffer and finalizedFBO[0] as proxy for the
system backbuffer. Why? Because otherwise the ordering of left and right
eye image/view gets swapped - not what we want.