Screen: Fixup another corner case of 'MirrorDisplayTo2ndOutputHead'.
The combination of a merged stereo mode (2 - 9) with Vulkan display will
trigger assignment of an actual OpenGL FBO to finalizedFBO's 0 and 1,
which makes sense for the general use case, but is detrimental to our
display mirroring approach as we need finalizedFBO[1] to represent the
system framebuffer during the final blit from bounce buffer texture to
slavewindow, not to an OpenGL FBO! To handle this edge case, use a little
hack: Temporarily assign fboid zero == System framebuffer for the duration
of the final blit, thren restore to the actula fboid binding.

This way the blit will go to the system fb / OpenGL backbuffer of the
slavewindow and merged stereomodes can also be mirrored if the Vulkan
backend is in use for the primary window.