Screen: Allow to disable vsync for slave windows selectively.
The new Screen('Openwindow', ..., specialflags) flag named
kPsychSkipSecondaryVsyncForFlip tells Screen that slave windows should
not use vsync'ed OpenGL bufferswaps. The one use case for this so far is
the PsychImaging task 'MirrorDisplayTo2ndOutputHead' for mirroring/cloning
the visual stimulus image presented to the subject with proper timing and
control into an associated slave window that just tries to show a copy at
roughly the same time. As the slave window is only meant for monitoring
feedback to the experimenter, quality aspects don't matter much, e.g.,
tearing should not be much of an issue. What does matter though is that
the presentation timing and framerate of animations in the main window
are not negatively affected by the mirrored stimuli, something that can
easily happen with vsync on if the control display has a lower / different
video refresh rate or non-synced refresh cycle with the main display.
Often the experimenter display will be cheaper / of lower or different
resolution / different model / different or lower video refresh rate, so
proper timing and performance across two vsync'ed displays is impossible.

Now we use the new kPsychSkipSecondaryVsyncForFlip flag to disable vsync
on the slave window, to trade lower quality/tearing for better timing.

For most (all current) other tasks involving a slave window, disabling
vsync doesn't make sense though: For dual-window stereo one wants both
images of a stereo-pair to be perfectly synchronized, which means vsync
on both windows and a software or hardware mechanism to sync the video
refresh cycles. Ditto for the 'DualPipeHDR' demo PsychImaging task, the
only (mockup of a) use case for dual-windows beyond stereo presentation,
where all outputs must be vsynced and in sync with each other.